perf: optimize render of BarWithTooltip with dumb component
